+++ transport/transport-runtime/src/androidTest/java/com/google/android/datatransport/runtime/UploaderIntegrationTest.java
@@ -17,6 +17,7 @@ package com.google.android.datatransport.runtime;
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -37,6 +38,7 @@ import com.google.android.datatransport.runtime.backends.TransportBackend;
 import com.google.android.datatransport.runtime.scheduling.jobscheduling.WorkScheduler;
 import com.google.android.datatransport.runtime.scheduling.persistence.EventStore;
 import com.google.android.datatransport.runtime.scheduling.persistence.PersistedEvent;
+import com.google.android.datatransport.runtime.synchronization.SynchronizationException;
 import java.util.Collections;
 import java.util.UUID;
 import org.junit.Before;
@@ -174,6 +176,35 @@ public class UploaderIntegrationTest {
         .send(eq(BackendRequest.create(Collections.singletonList(expectedEvent))));
     verify(spyScheduler, times(0)).schedule(any(), eq(2));
     assertThat(store.loadBatch(transportContext)).isEmpty();
+    assertThat(store.getNextCallTime(transportContext)).isEqualTo(0);
+  }
+
+  @Test
+  public void uploader_dbException_shouldReschedule() {
+    TransportRuntime runtime = TransportRuntime.getInstance();
+    EventStore store = component.getEventStore();
+    String mockBackendName = generateBackendName();
+
+    TransportContext transportContext =
+        TransportContext.builder()
+            .setBackendName(mockBackendName)
+            .setPriority(Priority.DEFAULT)
+            .build();
+    when(mockRegistry.get(mockBackendName)).thenReturn(mockBackend);
+    doThrow(new SynchronizationException("Error", null)).when(store).loadBatch(any());
+    TransportFactory factory = runtime.newFactory(mockBackendName);
+    Transport<String> transport =
+        factory.getTransport(testTransport, String.class, String::getBytes);
+    Event<String> stringEvent = Event.ofTelemetry("TelemetryData");
+    EventInternal expectedEvent =
+        EventInternal.builder()
+            .setEventMillis(3)
+            .setUptimeMillis(1)
+            .setTransportName(testTransport)
+            .setPayload("TelemetryData".getBytes())
+            .build();
+    transport.send(stringEvent);
+    verify(spyScheduler, times(1)).schedule(any(), eq(2));
     assertThat(store.getNextCallTime(transportContext)).isEqualTo(0);
   }
 }
